WebM Conversion
This is a command prompt script for Windows.  I'm afraid I don't know what the script is for Apple or Linux.

This script works only if the FFMPEG codec is installed.  Copy the script to a notepad and save as a batch file to the bin folder in the FFPEG install location.

Move video files to the same folder and run the batch file.  You should be okay from there.

Feel free to customize to your liking.